+++ app/lib/frontend/handlers/cache_control.dart
@@ -62,6 +62,20 @@ final class CacheControl {
     public: true,
   );
 
+  /// `Cache-Control` headers for package content pages, returning content that
+  /// is not updated frequently.
+  static const packageContentPage = CacheControl(
+    maxAge: Duration(minutes: 30),
+    public: true,
+  );
+
+  /// `Cache-Control` headers for package listing pages, returning content that
+  /// is may be updated frequently.
+  static const packageListingPage = CacheControl(
+    maxAge: Duration(minutes: 5),
+    public: true,
+  );
+
   /// `Cache-Control` headers for API end-points returning completion data for
   /// use in IDE integrations.
   static const completionData = CacheControl(